Fixing warnings

+++ skvideo/motion/block.py
@@ -12,9 +12,9 @@ def _costMAD(block1, block2):
 
 def _minCost(costs):
     h, w = costs.shape
-    mi = costs[((h-1)/2), ((w-1)/2)]
-    dy = ((h-1)/2)
-    dx = ((w-1)/2)
+    mi = costs[np.int((h-1)/2), np.int((w-1)/2)]
+    dy = np.int((h-1)/2)
+    dx = np.int((w-1)/2)
     #mi = 65535
     #dy = 0
     #dx = 0
@@ -53,7 +53,7 @@ def _DS(imgP, imgI, mbSize, p):
 
     h, w = imgP.shape
 
-    vectors = np.zeros((h / mbSize, w / mbSize, 2))
+    vectors = np.zeros((np.int(h / mbSize), np.int(w / mbSize), 2))
     costs = np.ones((9))*65537
 
     L = np.floor(np.log2(p + 1))
@@ -206,7 +206,7 @@ def _DS(imgP, imgI, mbSize, p):
             x += SDSP[point][0]
             y += SDSP[point][1]
 
-            vectors[i / mbSize, j / mbSize, :] = [x - j, y - i]
+            vectors[np.int(i / mbSize), np.int(j / mbSize), :] = [x - j, y - i]
 
             costs[:] = 65537
 
@@ -228,7 +228,7 @@ def _ARPS(imgP, imgI, mbSize, p):
 
     h, w = imgP.shape
 
-    vectors = np.zeros((h / mbSize, w / mbSize, 2))
+    vectors = np.zeros((np.int(h / mbSize), np.int(w / mbSize), 2))
     costs = np.ones((6))*65537
 
     SDSP = []
@@ -258,8 +258,8 @@ def _ARPS(imgP, imgI, mbSize, p):
                 stepSize = 2
                 maxIndex = 5
             else:
-                u = vectors[i / mbSize, j / mbSize - 1, 0]
-                v = vectors[i / mbSize, j / mbSize - 1, 1]
+                u = vectors[np.int(i / mbSize), np.int(j / mbSize) - 1, 0]
+                v = vectors[np.int(i / mbSize), np.int(j / mbSize) - 1, 1]
                 stepSize = np.int(np.max((np.abs(u), np.abs(v))))
 
                 if (((np.abs(u) == stepSize) and (np.abs(v) == 0)) or
@@ -340,7 +340,7 @@ def _ARPS(imgP, imgI, mbSize, p):
                     costs[:] = 65537
                     costs[2] = cost
 
-            vectors[i / mbSize, j / mbSize, :] = [x - j, y - i]
+            vectors[np.int(i / mbSize), np.int(j / mbSize), :] = [x - j, y - i]
 
             costs[:] = 65537
 
@@ -364,7 +364,7 @@ def _SE3SS(imgP, imgI, mbSize, p):
 
     h, w = imgP.shape
 
-    vectors = np.zeros((h / mbSize, w / mbSize, 2))
+    vectors = np.zeros((np.int(h / mbSize), np.int(w / mbSize), 2))
     L = np.floor(np.log2(p + 1))
     stepMax = 2**(L - 1)
 
@@ -490,7 +490,7 @@ def _SE3SS(imgP, imgI, mbSize, p):
                 costs[:] = 65537
                 stepSize /= 2
 
-            vectors[i / mbSize, j / mbSize, :] = [y - i, x - j]
+            vectors[np.int(i / mbSize), np.int(j / mbSize), :] = [y - i, x - j]
 
             costs[:] = 65537
 
@@ -517,7 +517,7 @@ def _N3SS(imgP, imgI, mbSize, p):
     imgP = imgP[:h, :w]
     imgI = imgI[:h, :w]
 
-    vectors = np.zeros((h / mbSize, w / mbSize, 2), dtype=np.float32)
+    vectors = np.zeros((np.int(h / mbSize), np.int(w / mbSize), 2), dtype=np.float32)
 
     costs = np.ones((3, 3), dtype=np.float32)*65537
 
@@ -546,8 +546,8 @@ def _N3SS(imgP, imgI, mbSize, p):
                        (refBlkHor < 0) or
                        (refBlkHor + mbSize > w)):
                             continue
-                    costRow = m / stepSize + 1
-                    costCol = n / stepSize + 1
+                    costRow = np.int(m / stepSize) + 1
+                    costCol = np.int(n / stepSize) + 1
                     if ((costRow == 1) and (costCol == 1)):
                         continue
                     costs[costRow, costCol] = _costMAD(imgP[i:i + mbSize, j:j + mbSize], imgI[refBlkVer:refBlkVer + mbSize, refBlkHor:refBlkHor + mbSize])
@@ -612,8 +612,8 @@ def _N3SS(imgP, imgI, mbSize, p):
                             (refBlkHor >= j - 1) and
                             (refBlkHor <= j + 1)):
                                 continue
-                        costRow = m/stepSize + 1
-                        costCol = n/stepSize + 1
+                        costRow = np.int(m/stepSize) + 1
+                        costCol = np.int(n/stepSize) + 1
                         if ((costRow == 1) and (costCol == 1)):
                             continue
                         costs[costRow, costCol] = _costMAD(imgP[i:i + mbSize, j:j + mbSize], imgI[refBlkVer:refBlkVer + mbSize, refBlkHor:refBlkHor + mbSize])
@@ -653,7 +653,7 @@ def _N3SS(imgP, imgI, mbSize, p):
                     costs[1, 1] = costs[dy, dx]
 
 
-            vectors[i / mbSize, j / mbSize, :] = [y - i, x - j]
+            vectors[np.int(i / mbSize), np.int(j / mbSize), :] = [y - i, x - j]
 
             costs[:, :] = 65537
 
@@ -664,7 +664,7 @@ def _N3SS(imgP, imgI, mbSize, p):
 def _3SS(imgP, imgI, mbSize, p):
     h, w = imgP.shape
 
-    vectors = np.zeros((h / mbSize, w / mbSize, 2))
+    vectors = np.zeros((np.int(h / mbSize), np.int(w / mbSize), 2))
     costs = np.ones((3, 3), dtype=np.float32)*65537
 
     computations = 0
@@ -692,8 +692,8 @@ def _3SS(imgP, imgI, mbSize, p):
                            (refBlkHor < 0) or
                            (refBlkHor + mbSize > w)):
                                 continue
-                        costRow = m/stepSize + 1
-                        costCol = n/stepSize + 1
+                        costRow = np.int(m/stepSize) + 1
+                        costCol = np.int(n/stepSize) + 1
                         if ((costRow == 1) and (costCol == 1)):
                             continue
                         costs[costRow, costCol] = _costMAD(imgP[i:i + mbSize, j:j + mbSize], imgI[refBlkVer:refBlkVer + mbSize, refBlkHor:refBlkHor + mbSize])
@@ -704,7 +704,7 @@ def _3SS(imgP, imgI, mbSize, p):
 
                 stepSize /= 2
                 costs[1, 1] = costs[dy, dx]
-            vectors[i / mbSize, j / mbSize, :] = [y - i, x - j]
+            vectors[np.int(i / mbSize), np.int(j / mbSize), :] = [y - i, x - j]
 
             costs[:, :] = 65537
 
@@ -724,7 +724,7 @@ def _4SS(imgP, imgI, mbSize, p):
     #   SS4computations: The average number of points searched for a macroblock
     h, w = imgP.shape
 
-    vectors = np.zeros((h / mbSize, w / mbSize, 2))
+    vectors = np.zeros((np.int(h / mbSize), np.int(w / mbSize), 2))
     costs = np.ones((3, 3), dtype=np.float32)*65537
 
     computations = 0
@@ -780,8 +780,8 @@ def _4SS(imgP, imgI, mbSize, p):
                             (refBlkVer >= yLast + 2)):
                             continue
 
-                        costRow = m/2 + 1
-                        costCol = n/2 + 1
+                        costRow = np.int(m/2) + 1
+                        costCol = np.int(n/2) + 1
 
                         if (costRow == 1 and costCol == 1):
                             continue
@@ -824,7 +824,7 @@ def _4SS(imgP, imgI, mbSize, p):
             x += dx - 1
             y += dy - 1
 
-            vectors[i / mbSize, j / mbSize, :] = [y - i, x - j]
+            vectors[np.int(i / mbSize), np.int(j / mbSize), :] = [y - i, x - j]
 
             costs[:, :] = 65537
 
@@ -835,7 +835,7 @@ def _4SS(imgP, imgI, mbSize, p):
 def _ES(imgP, imgI, mbSize, p):
     h, w = imgP.shape
 
-    vectors = np.zeros((h / mbSize, w / mbSize, 2), dtype=np.float32)
+    vectors = np.zeros((np.int(h / mbSize), np.int(w / mbSize), 2), dtype=np.float32)
     costs = np.ones((2 * p + 1, 2 * p + 1), dtype=np.float32)*65537
 
     # we start off from the top left of the image
@@ -878,7 +878,7 @@ def _ES(imgP, imgI, mbSize, p):
             # Now we find the vector where the cost is minimum
             # and store it ... this is what will be passed back.
             dx, dy, mi = _minCost(costs)  # finds which macroblock in imgI gave us min Cost
-            vectors[i / mbSize, j / mbSize, :] = [dy - p, dx - p]
+            vectors[np.int(i / mbSize), np.int(j / mbSize), :] = [dy - p, dx - p]
 
             costs[:, :] = 65537
 
@@ -947,7 +947,7 @@ def blockMotion(videodata, method='DS', mbSize=8, p=2, **plugin_args):
     # luminance is 1 channel, so flatten for computation
     luminancedata = luminancedata.reshape((numFrames, height, width))
 
-    motionData = np.zeros((numFrames - 1, height / mbSize, width / mbSize, 2), np.int8)
+    motionData = np.zeros((numFrames - 1, np.int(height / mbSize), np.int(width / mbSize), 2), np.int8)
 
     if method == "ES":
         for i in range(numFrames - 1):
